Replace Travis CI config with GitHub Actions
This commit removes the Travis config and replaces
it with a GitHub actions workflow, which is standard
for dxw repositories.

See:
    https://blog.travis-ci.com/2020-11-02-travis-ci-new-billing

The new configuration is based on Terraform 0.12.0, following
the changes in 114d53e.

The validation command also sets a default AWS region, to work
around validation failures caused by this issue:
https://github.com/hashicorp/terraform/issues/21408